requirements.txt
# Waste Classification System - Requirements
# Production-ready dependencies for ML pipeline, API, and frontend

# ==============================================================================
# Core ML Libraries
# ==============================================================================
tensorflow>=2.15.0
numpy
pandas>=2.0.0
scikit-learn>=1.3.0

# Model Optimization (optional - install separately if needed for pruning)
# For Python 3.12+, tensorflow-model-optimization may have compatibility issues
# pip install tensorflow-model-optimization  # Run separately if needed

# ==============================================================================
# Image Processing
# ==============================================================================
opencv-python>=4.7.0
Pillow>=9.4.0

# ==============================================================================
# Web Framework - Backend
# ==============================================================================
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0

# ==============================================================================
# Web Framework - Frontend
# ==============================================================================
streamlit>=1.25.0

# ==============================================================================
# Visualization
# ==============================================================================
matplotlib>=3.7.0
seaborn>=0.12.0

# ==============================================================================
# HTTP Client
# ==============================================================================
requests>=2.28.0
aiohttp>=3.8.0

# ==============================================================================
# Utilities
# ==============================================================================
python-dotenv>=1.0.0

# ==============================================================================
# Development & Notebooks
# ==============================================================================
jupyter>=1.0.0
ipywidgets>=8.0.0
